move empty state icon out of subscriptions widget
